+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [2.2.0](https://github.com/netgrif/petriflow.js/releases/tag/v2.2.0) (2024-07-26)
+
+### Changed
+- [PF-71] Sort collections before export
+- [PF-72] Tags
+
+
+## [2.1.0](https://github.com/netgrif/petriflow.js/releases/tag/v2.1.0) (2024-07-02)
+
+### Changed
+- [PF-58] Action indentation
+
+
 ## [2.0.0](https://github.com/netgrif/petriflow.js/releases/tag/v2.0.0) (2023-11-05)
 
 ### Added